w3c-reports/12092014/efl/runned-tests/IndexedDB/idbtransaction-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: indexedDB

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: indexedDB
